moved Atomic.h to new location, moved code to CUDAUtils.h
