+++ src/Open3D/Container/CUDAUtils.h
@@ -65,6 +65,22 @@ inline void __OPEN3D_CUDA_CHECK(cudaError_t err,
                           cudaGetErrorString(err));
     }
 }
+
+/// Returns the texture alignment in bytes for the current device.
+inline int GetCUDACurrentDeviceTextureAlignment() {
+    int device = 0;
+    cudaError_t err = cudaGetDevice(&device);
+    if (err != cudaSuccess) {
+      utility::LogError("GetCUDACurrentDeviceTextureAlignment(): cudaGetDevice failed with {}", cudaGetErrorString(err));
+    }
+
+    int value = 0;
+    err = cudaDeviceGetAttribute(&value, cudaDevAttrTextureAlignment, device);
+    if (err != cudaSuccess) {
+      utility::LogError("GetCUDACurrentDeviceTextureAlignment(): cudaDeviceGetAttribute failed with {}", cudaGetErrorString(err));
+    }
+    return value;
+}
 #endif
 
 }  // namespace open3d
